requirements.txt
attrs==19.3.0
better-profanity==0.6.1
click==7.1.2
coverage==5.2.1
Flask==1.1.2
flask-paginate==0.7.1
Flask-WTF==0.14.2
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
numpy==1.19.2
packaging==20.4
pandas==1.1.3
password-validator==1.0
pluggy==0.13.1
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
wcwidth==0.2.4
Werkzeug==0.16.0
WTForms==2.3.3
zipp==3.3.0

SQLAlchemy~=1.3.20